backend/requirements.txt
Django
djangorestframework
tensorflow
Pillow
djangorestframework-simplejwt
django-cors-headers
gunicorn
